+++ lib/utils/chart_content.php
@@ -7,7 +7,7 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
 
     $next_theme_id = $chart->getTheme();
 
-    $locale = $user->getLanguage();
+    $locale = DatawrapperSession::getLanguage();
     $themeLocale = null;
 
     while (!empty($next_theme_id)) {
@@ -22,11 +22,6 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
         }
         $next_theme_id = $theme['extends'];
     }
-    // theme locale overrides user locale
-    if (!empty($themeLocale)) $locale = $themeLocale;
-    // per-chart locale overrides theme locale
-    // $chartLocale = $chart->getLanguage();
-    // if (!empty($chartLocale)) $locale = $chartLocale;
 
     $abs = 'http://' . $GLOBALS['dw_config']['domain'];
 
@@ -36,7 +31,7 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
         $base_js = array(
             '//assets-datawrapper.s3.amazonaws.com/globalize.min.js',
             '//cdnjs.cloudflare.com/ajax/libs/underscore.js/1.4.2/underscore-min.js',
-            '//cdnjs.cloudflare.com/ajax/libs/jquery/1.8.2/jquery.min.js'
+            '//cdnjs.cloudflare.com/ajax/libs/jquery/1.9.1/jquery.min.js'
         );
     } else {
         // use local assets
@@ -47,11 +42,17 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
         );
     }
 
+    if (substr($locale, 0, 2) != 'en') {
+        $base_js[] = $abs . '/static/vendor/globalize/cultures/globalize.culture.' . str_replace('_', '-', $locale) . '.js';
+    }
+
     $vis_js = array();
     $vis_css = array();
     $next_vis_id = $chart->getType();
 
     $vis_libs = array();
+
+    $vis_locale = array();  // visualizations may define localized strings, e.g. "other"
 
     while (!empty($next_vis_id)) {
         $vis = get_visualization_meta($next_vis_id, $path);
@@ -59,6 +60,11 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
         if (!empty($vis['libraries'])) {
             foreach ($vis['libraries'] as $url) {
                 $vis_libs[] = '/static/vendor/' . $url;
+            }
+        }
+        if (!empty($vis['locale'])) {
+            foreach ($vis['locale'] as $term => $translations) {
+                if (!isset($vis_locale[$term])) $vis_locale[$term] = $translations;
             }
         }
         $vjs[] = '/static/visualizations/' . $vis['id'] . '/' . $vis['id'] . '.js';
@@ -72,6 +78,7 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
     $styles = array_merge($vis_css, array_reverse($theme_css));
 
     $the_vis = get_visualization_meta($chart->getType(), $path);
+    $the_vis['locale'] = $vis_locale;
     $the_theme = get_theme_meta($chart->getTheme(), $path);
 
     if ($minified) {
@@ -109,6 +116,7 @@ function get_chart_content($chart, $user, $minified = false, $path = '') {
         'chartData' => $chart->loadData(),
         'chart' => $chart,
         'chartLocale' => str_replace('_', '-', $locale),
+        'lang' => strtolower(substr($locale, 0, 2)),
         'metricPrefix' => get_metric_prefix($locale),
         'theme' => $the_theme,
         'visualization' => $the_vis,