+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 ---------
 
+### 1.2.6
+* visualizations now may define [localized strings](https://github.com/datawrapper/datawrapper/wiki/Visualization-Meta.json-Properties#locale)
+* fixed bug with localized numbers in charts
+
 ### 1.2.5
 * added percentage stacking mode to stacked column charts
 * improved session garbage collection